Move `ceil()` and `floor()` tests to builtinTest.py
